requirements.txt
Flask
mysql-connector-python
langchain-google-genai
openai
langchain
